requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1
django-heroku==0.3.1
djangorestframework==3.14.0
frozenlist==1.3.1
github==1.2.7
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
multidict==6.0.2
Pillow==9.2.0
psycopg2==2.9.4
python-dateutil==1.5
pytz==2022.5
requests==2.28.1
sqlparse==0.4.2
typing_extensions==4.4.0
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
yarl==1.8.1
